+++ utoipa/src/openapi.rs
@@ -163,11 +163,11 @@ impl OpenApi {
         serde_json::to_string_pretty(self)
     }
 
-    /// Converts this [`OpenApi`] to YAML String. This method essentially calls [`serde_yaml::to_string`] method.
+    /// Converts this [`OpenApi`] to YAML String. This method essentially calls [`serde_yml::to_string`] method.
     #[cfg(feature = "yaml")]
     #[cfg_attr(doc_cfg, doc(cfg(feature = "yaml")))]
-    pub fn to_yaml(&self) -> Result<String, serde_yaml::Error> {
-        serde_yaml::to_string(self)
+    pub fn to_yaml(&self) -> Result<String, serde_yml::Error> {
+        serde_yml::to_string(self)
     }
 
     /// Merge `other` [`OpenApi`] moving `self` and returning combined [`OpenApi`].